Split model to its own file and fixed imports
There is a bit of an oddity where we're asking the window if a block is a cloud
instead of asking the model, however fixing that is outside of the scope of this
change.
